+++ docs/source/changelog.rst
@@ -12,7 +12,7 @@ API Breaking Changes
 Enhancements
 ------------
 
-- Added a new meta-estimator :class:`dask_ml.wrappers.Blockwise` for wrapping any estimator with a `partial_fit` method (). See :ref:`incremental.blockwise-metaestimator` for more.
+- Added a new meta-estimator :class:`dask_ml.wrappers.Incremental` for wrapping any estimator with a `partial_fit` method (). See :ref:`incremental.blockwise-metaestimator` for more.
 
 Version 0.5.0
 ~~~~~~~~~~~~~